third_party/WebKit/LayoutTests/external/wpt/web-animations/animation-model/keyframe-effects/effect-value-context-expected.txt

This is a testharness.js-based test.
PASS Effect values reflect changes to font-size on element 
PASS Effect values reflect changes to font-size on parent element 
PASS Effect values reflect changes to font-size when computed style is not immediately flushed 
PASS Effect values reflect changes to font-size from reparenting 
FAIL Effect values reflect changes to target element assert_equals: Effect value after updating target element expected "300px" but got "0px"
Harness: the test ran to completion.
